Conflicts:
	src/icrar/leap-accelerate/CMakeLists.txt
	src/icrar/leap-accelerate/algorithm/casa/PhaseMatrixFunction.cc
	src/icrar/leap-accelerate/algorithm/casa/PhaseRotate.cc
	src/icrar/leap-accelerate/algorithm/cuda/PhaseRotate.cu
	src/icrar/leap-accelerate/common/vector_extensions.h
	src/icrar/leap-accelerate/exception/exception.h
	src/icrar/leap-accelerate/model/casa/CalibrateResult.h
	src/icrar/leap-accelerate/model/casa/Integration.cc
	src/icrar/leap-accelerate/model/casa/Integration.h
	src/icrar/leap-accelerate/model/casa/MetaData.cc
	src/icrar/leap-accelerate/model/casa/MetaData.h
	src/icrar/leap-accelerate/model/cpu/CalibrateResult.cc
	src/icrar/leap-accelerate/model/cpu/MetaData.h
	src/icrar/leap-accelerate/ms/MeasurementSet.h

+++ src/icrar/leap-accelerate/math/math_conversion.h
@@ -71,7 +71,7 @@ namespace icrar
     template<typename T>
     casacore::Matrix<T> ConvertMatrix(const Eigen::Matrix<T, Eigen::Dynamic, Eigen::Dynamic>& value)
     {
-        return casacore::Matrix<T>(casacore::IPosition(2, (int)value.rows(), (int)value.cols()), value.data());
+        return casacore::Matrix<T>(casacore::IPosition(2, value.rows(), value.cols()), value.data());
     }
 
     template<typename T>
@@ -129,4 +129,4 @@ namespace icrar
 
     casacore::MVDirection ToCasaDirection(const icrar::MVDirection& value);
     std::vector<casacore::MVDirection> ToCasaDirectionVector(const std::vector<icrar::MVDirection>& value);
-}
+} // namespace icrar
